[mob] Remove embedding store
